+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Renamed object from `ComplianceData` to `BatchComplianceData`. If you are using the object name `ComplianceData` explicitly, modify it to `BatchComplianceData`. ([#462](https://github.com/twitter-dart/twitter-api-v2/issues/462))
 - Deprecated `connectVolumeStream` and added `connectSampleStream` instead. Please replace method name from `connectVolumeStream` to `connectSampleStream`. ([#479](https://github.com/twitter-dart/twitter-api-v2/issues/479))
+- Supported `connectSample10Stream` in Tweets Service. This feature is available to Enterprise levels of access. Also, this is commonly referred to as the **Decahose**. ([#473](https://github.com/twitter-dart/twitter-api-v2/issues/473))
 
 ## v4.1.3